requirements.txt
fastapi
uvicorn
pymupdf
sentence-transformers
faiss-cpu
transformers
llama-cpp-python
python-multipart